Struct re_types::archetypes::Asset3D

pub struct Asset3D {
    pub blob: Blob,
    pub media_type: Option<MediaType>,
    pub transform: Option<OutOfTreeTransform3D>,
}

Archetype: A prepacked 3D asset (.gltf, .glb, .obj, etc.).

See also Mesh3D.

Examples

Simple 3D asset

//! Log a simple 3D asset.

use rerun::external::anyhow;

fn main() -> anyhow::Result<()> {
    let args = std::env::args().collect::<Vec<_>>();
    let Some(path) = args.get(1) else {
        anyhow::bail!("Usage: {} <path_to_asset.[gltf|glb]>", args[0]);
    };

    let (rec, storage) =
        rerun::RecordingStreamBuilder::new("rerun_example_asset3d_simple").memory()?;

    rec.log_timeless("world", &rerun::ViewCoordinates::RIGHT_HAND_Z_UP)?; // Set an up-axis
    rec.log("world/asset", &rerun::Asset3D::from_file(path)?)?;

    rerun::native_viewer::show(storage.take())?;
    Ok(())
}

3D asset with out-of-tree transform

//! Log a simple 3D asset with an out-of-tree transform which will not affect its children.

use rerun::{
    demo_util::grid,
    external::{anyhow, glam},
};

fn main() -> anyhow::Result<()> {
    let args = std::env::args().collect::<Vec<_>>();
    let Some(path) = args.get(1) else {
        anyhow::bail!("Usage: {} <path_to_asset.[gltf|glb]>", args[0]);
    };

    let (rec, storage) =
        rerun::RecordingStreamBuilder::new("rerun_example_asset3d_out_of_tree").memory()?;

    rec.log_timeless("world", &rerun::ViewCoordinates::RIGHT_HAND_Z_UP)?; // Set an up-axis

    rec.set_time_sequence("frame", 0);
    rec.log("world/asset", &rerun::Asset3D::from_file(path)?)?;
    // Those points will not be affected by their parent's out-of-tree transform!
    rec.log(
        "world/asset/points",
        &rerun::Points3D::new(grid(glam::Vec3::splat(-10.0), glam::Vec3::splat(10.0), 10)),
    )?;

    for i in 1..20 {
        rec.set_time_sequence("frame", i);

        // Modify the asset's out-of-tree transform: this will not affect its children (i.e. the points)!
        let translation =
            rerun::TranslationRotationScale3D::from_translation([0.0, 0.0, i as f32 - 10.0]);
        rec.log_component_batches(
            "world/asset",
            false,
            [&rerun::OutOfTreeTransform3D::from(translation) as _],
        )?;
    }

    rerun::native_viewer::show(storage.take())?;
    Ok(())
}

Fields
blob: Blob

The asset’s bytes.

media_type: Option<MediaType>

The Media Type of the asset.

For instance:

model/gltf-binary

model/obj

If omitted, the viewer will try to guess from the data blob. If it cannot guess, it won’t be able to render the asset.

transform: Option<OutOfTreeTransform3D>

An out-of-tree transform.

Applies a transformation to the asset itself without impacting its children.

Implementations

impl Asset3D

pub const NUM_COMPONENTS: usize = 5usize

impl Asset3D

pub fn new(blob: impl Into<Blob>) -> Self

pub fn with_media_type(self, media_type: impl Into<MediaType>) -> Self

pub fn with_transform(self, transform: impl Into<OutOfTreeTransform3D>) -> Self

impl Asset3D

pub fn from_file(path: impl AsRef<Path>) -> Result<Self>

Creates a new Asset3D from the file contents at path.

The MediaType will first be guessed from the file extension, then from the file contents if needed.

If no MediaType can be guessed at the moment, the Rerun Viewer will try to guess one from the data at render-time. If it can’t, rendering will fail with an error.

pub fn from_bytes( bytes: impl AsRef<[u8]>, media_type: Option<impl Into<MediaType>> ) -> Self

Creates a new Asset3D from the given bytes.

The MediaType will be guessed from magic bytes in the data.

If no MediaType can be guessed at the moment, the Rerun Viewer will try to guess one from the data at render-time. If it can’t, rendering will fail with an error.