Struct re_types::archetypes::DisconnectedSpace

pub struct DisconnectedSpace {
    pub disconnected_space: DisconnectedSpace,
}

Archetype: Specifies that the entity path at which this is logged is disconnected from its parent.

This is useful for specifying that a subgraph is independent of the rest of the scene.

If a transform or pinhole is logged on the same path, this archetype’s components will be ignored.

Example

disconnected_space:

//! Disconnect two spaces.

fn main() -> Result<(), Box<dyn std::error::Error>> {
    let (rec, storage) =
        rerun::RecordingStreamBuilder::new("rerun_example_disconnected_space").memory()?;

    // These two points can be projected into the same space..
    rec.log(
        "world/room1/point",
        &rerun::Points3D::new([(0.0, 0.0, 0.0)]),
    )?;
    rec.log(
        "world/room2/point",
        &rerun::Points3D::new([(1.0, 1.0, 1.0)]),
    )?;

    // ..but this one lives in a completely separate space!
    rec.log("world/wormhole", &rerun::DisconnectedSpace::new(true))?;
    rec.log(
        "world/wormhole/point",
        &rerun::Points3D::new([(2.0, 2.0, 2.0)]),
    )?;

    rerun::native_viewer::show(storage.take())?;
    Ok(())
}

Fields
disconnected_space: DisconnectedSpace
Implementations

impl DisconnectedSpace

pub const NUM_COMPONENTS: usize = 3usize

impl DisconnectedSpace

pub fn new(disconnected_space: impl Into<DisconnectedSpace>) -> Self