Struct re_types::blueprint::archetypes::space_view_blueprint::SpaceViewBlueprint

pub struct SpaceViewBlueprint {
    pub class_identifier: SpaceViewClass,
    pub display_name: Option<Name>,
    pub space_origin: Option<SpaceViewOrigin>,
    pub entities_determined_by_user: Option<EntitiesDeterminedByUser>,
    pub contents: Option<IncludedQueries>,
    pub visible: Option<Visible>,
}

Archetype: The top-level description of the Viewport.

Fields
class_identifier: SpaceViewClass

The class of the view.

display_name: Option<Name>

The name of the view.

space_origin: Option<SpaceViewOrigin>

The “anchor point” of this space view.

The transform at this path forms the reference point for all scene->world transforms in this space view. I.e. the position of this entity path in space forms the origin of the coordinate system in this space view. Furthermore, this is the primary indicator for heuristics on what entities we show in this space view.

entities_determined_by_user: Option<EntitiesDeterminedByUser>

True if the user is has added entities themselves. False otherwise.

contents: Option<IncludedQueries>

BlueprintIds of the DataQuerys that make up this SpaceView.

It determines which entities are part of the spaceview.

visible: Option<Visible>

Whether this space view is visible.

Defaults to true if not specified.

Implementations

impl SpaceViewBlueprint

pub const NUM_COMPONENTS: usize = 8usize

impl SpaceViewBlueprint

pub fn new(class_identifier: impl Into<SpaceViewClass>) -> Self

pub fn with_display_name(self, display_name: impl Into<Name>) -> Self

pub fn with_space_origin(self, space_origin: impl Into<SpaceViewOrigin>) -> Self

pub fn with_entities_determined_by_user( self, entities_determined_by_user: impl Into<EntitiesDeterminedByUser> ) -> Self

pub fn with_contents(self, contents: impl Into<IncludedQueries>) -> Self

pub fn with_visible(self, visible: impl Into<Visible>) -> Self