Module re_types::blueprint::components

The blueprint-specific components.

Modules

active_tab 🔒

apply_latest_at 🔒

background_kind 🔒

column_share 🔒

component_column_selector 🔒

component_column_selector_ext 🔒

corner2d 🔒

corner2d_ext 🔒

filter_by_range 🔒

filter_by_range_ext 🔒

filter_is_not_null 🔒

filter_is_not_null_ext 🔒

included_content 🔒

interactive 🔒

interactive_ext 🔒

lock_range_during_zoom 🔒

panel_state 🔒

panel_state_ext 🔒

query_expression 🔒

row_share 🔒

selected_columns 🔒

space_view_class 🔒

space_view_class_ext 🔒

space_view_origin 🔒

space_view_origin_ext 🔒

tensor_dimension_index_slider 🔒

tensor_dimension_index_slider_ext 🔒

timeline_name 🔒

timeline_name_ext 🔒

view_fit 🔒

viewer_recommendation_hash 🔒

viewer_recommendation_hash_ext 🔒

visible 🔒

visible_ext 🔒

visible_time_range 🔒

visual_bounds2d 🔒

visual_bounds2d_ext 🔒

Structs

ActiveTab

Component: The active tab in a tabbed container.

ApplyLatestAt

Component: Whether empty cells in a dataframe should be filled with a latest-at query.

ColumnShare

Component: The layout share of a column in the container.

ComponentColumnSelector

Component: Describe a component column to be selected in the dataframe view.

FilterByRange

Component: Configuration for a filter-by-range feature of the dataframe view.

FilterIsNotNull

Component: Configuration for the filter is not null feature of the dataframe view.

IncludedContent

Component: All the contents in the container.

Interactive

Component: Whether the entity can be interacted with.

LockRangeDuringZoom

Component: Indicate whether the range should be locked when zooming in on the data.

QueryExpression

Component: An individual query expression used to filter a set of datatypes::EntityPaths.

RowShare

Component: The layout share of a row in the container.

SelectedColumns

Component: Describe a component column to be selected in the dataframe view.

SpaceViewClass

Component: The class identifier of view, e.g. "2D", "TextLog", ….

SpaceViewOrigin

Component: The origin of a SpaceView.

TensorDimensionIndexSlider

Component: Show a slider for the index of some dimension of a slider.

TimelineName

Component: A timeline identified by its name.

ViewerRecommendationHash

Component: Hash of a viewer recommendation.

Visible

Component: Whether the container, view, entity or instance is currently visible.

VisibleTimeRange

Component: The range of values on a given timeline that will be included in a view’s query.

VisualBounds2D

Component: Visual bounds in 2D space used for Spatial2DView.

Enums

BackgroundKind

Component: The type of the background in a view.

Corner2D

Component: One of four 2D corners, typically used to align objects.

PanelState

Component: Tri-state for panel controls.

ViewFit

Component: Determines whether an image or texture should be scaled to fit the viewport.