re_types::blueprint
Module datatypes

The blueprint-specific datatypes.

Modules

component_column_selector 🔒

component_column_selector_ext 🔒

filter_by_range 🔒

filter_is_not_null 🔒

selected_columns 🔒

tensor_dimension_index_slider 🔒

Structs

ComponentColumnSelector

Datatype: Describe a component column to be selected in the dataframe view.

FilterByRange

Datatype: Configuration for the filter-by-range feature of the dataframe view.

FilterIsNotNull

Datatype: Configuration for the filter is not null feature of the dataframe view.

SelectedColumns

Datatype: List of selected columns in a dataframe.

TensorDimensionIndexSlider

Datatype: Defines a slider for the index of some dimension.