Module re_types::components

Modules

annotation_context ğŸ”’

class_id ğŸ”’

class_id_ext ğŸ”’

color ğŸ”’

color_ext ğŸ”’

disconnected_space ğŸ”’

disconnected_space_ext ğŸ”’

draw_order ğŸ”’

draw_order_ext ğŸ”’

instance_key ğŸ”’

instance_key_ext ğŸ”’

keypoint_id ğŸ”’

keypoint_id_ext ğŸ”’

label ğŸ”’

label_ext ğŸ”’

line_strip2d ğŸ”’

line_strip2d_ext ğŸ”’

line_strip3d ğŸ”’

line_strip3d_ext ğŸ”’

origin3d ğŸ”’

point2d ğŸ”’

point2d_ext ğŸ”’

point3d ğŸ”’

point3d_ext ğŸ”’

radius ğŸ”’

radius_ext ğŸ”’

transform3d ğŸ”’

transform3d_ext ğŸ”’

vector3d ğŸ”’

Structs

AnnotationContext

The AnnotationContext provides additional information on how to display entities.

ClassId

A 16-bit ID representing a type of semantic class.

Color

An RGBA color tuple with unmultiplied/separate alpha, in sRGB gamma space with linear alpha.

DisconnectedSpace

Specifies that the entity path at which this is logged is disconnected from its parent.

DrawOrder

Draw order used for the display order of 2D elements.

InstanceKey

A unique numeric identifier for each individual instance within a batch.

KeypointId

A 16-bit ID representing a type of semantic keypoint within a class.

Label

A String label component.

LineStrip2D

A line strip in 2D space.

LineStrip3D

A line strip in 3D space.

Origin3D

A point of origin in 3D space.

Point2D

A point in 2D space.

Point3D

A point in 3D space.

Radius

A Radius component.

Transform3D

An affine transform between two 3D spaces, represented in a given direction.

Vector3D

A vector in 3D space.