Struct re_types::components::draw_order::DrawOrder

#[repr(transparent)]
pub struct DrawOrder(pub f32);

Component: Draw order used for the display order of 2D elements.

Higher values are drawn on top of lower values. An entity can have only a single draw order component. Within an entity draw order is governed by the order of the components.

Draw order for entities with the same draw order is generally undefined.

Tuple Fields
0: f32
Implementations

impl DrawOrder

pub const DEFAULT_IMAGE: DrawOrder = _

Draw order used for images if no draw order was specified.

pub const DEFAULT_BOX2D: DrawOrder = _

Draw order used for 2D boxes if no draw order was specified.

pub const DEFAULT_LINES2D: DrawOrder = _

Draw order used for 2D lines if no draw order was specified.

pub const DEFAULT_POINTS2D: DrawOrder = _

Draw order used for 2D points if no draw order was specified.