Struct re_types::components::ViewCoordinates

#[repr(transparent)]
pub struct ViewCoordinates(pub [u8; 3]);

Component: How we interpret the coordinate system of an entity/space.

For instance: What is “up”? What does the Z axis mean? Is this right-handed or left-handed?

The three coordinates are always ordered as [x, y, z].

For example [Right, Down, Forward] means that the X axis points to the right, the Y axis points down, and the Z axis points forward.

The following constants are used to represent the different directions. Up = 1 Down = 2 Right = 3 Left = 4 Forward = 5 Back = 6

Tuple Fields
0: [u8; 3]

The directions of the [x, y, z] axes.

Implementations

impl ViewCoordinates

pub const fn new(x: ViewDir, y: ViewDir, z: ViewDir) -> Self

Construct a new ViewCoordinates from an array of ViewDirs.

pub fn from_up_and_handedness(up: SignedAxis3, handedness: Handedness) -> Self

Choses a coordinate system based on just an up-axis.

pub fn sanity_check(&self) -> Result<(), String>

Returns an error if this does not span all three dimensions.

pub fn up(&self) -> Option<SignedAxis3>

pub fn right(&self) -> Option<SignedAxis3>

pub fn forward(&self) -> Option<SignedAxis3>

pub fn describe_short(&self) -> String

pub fn describe(&self) -> String

pub fn from_other(&self, other: &Self) -> Mat3

Returns a matrix that transforms from another coordinate system to this (self) one.

pub fn to_rdf(&self) -> Mat3

Returns a matrix that transforms this coordinate system to RDF.

(RDF: X=Right, Y=Down, Z=Forward)

pub fn from_rdf(&self) -> Mat3

Returns a matrix that transforms from RDF to this coordinate system.

(RDF: X=Right, Y=Down, Z=Forward)

pub fn to_rub(&self) -> Mat3

Returns a matrix that transforms this coordinate system to RUB.

(RUB: X=Right, Y=Up, Z=Back)

pub fn from_rub(&self) -> Mat3

Returns a matrix that transforms from RUB to this coordinate system.

(RUB: X=Right, Y=Up, Z=Back)

pub fn from_rub_quat(&self) -> Result<Quat, String>

Returns a quaternion that rotates from RUB to this coordinate system.

Errors if the coordinate system is left-handed or degenerate.

(RUB: X=Right, Y=Up, Z=Back)

pub fn handedness(&self) -> Result<Handedness, String>

impl ViewCoordinates

pub const ULF: Self = _

pub const UFL: Self = _

pub const LUF: Self = _

pub const LFU: Self = _

pub const FUL: Self = _

pub const FLU: Self = _

pub const ULB: Self = _

pub const UBL: Self = _

pub const LUB: Self = _

pub const LBU: Self = _

pub const BUL: Self = _

pub const BLU: Self = _

pub const URF: Self = _

pub const UFR: Self = _

pub const RUF: Self = _

pub const RFU: Self = _

pub const FUR: Self = _

pub const FRU: Self = _

pub const URB: Self = _

pub const UBR: Self = _

pub const RUB: Self = _

pub const RBU: Self = _

pub const BUR: Self = _

pub const BRU: Self = _

pub const DLF: Self = _

pub const DFL: Self = _

pub const LDF: Self = _

pub const LFD: Self = _

pub const FDL: Self = _

pub const FLD: Self = _

pub const DLB: Self = _

pub const DBL: Self = _

pub const LDB: Self = _

pub const LBD: Self = _

pub const BDL: Self = _

pub const BLD: Self = _

pub const DRF: Self = _

pub const DFR: Self = _

pub const RDF: Self = _

pub const RFD: Self = _

pub const FDR: Self = _

pub const FRD: Self = _

pub const DRB: Self = _

pub const DBR: Self = _

pub const RDB: Self = _

pub const RBD: Self = _

pub const BDR: Self = _

pub const BRD: Self = _

pub const RIGHT_HAND_X_UP: Self = _

pub const RIGHT_HAND_X_DOWN: Self = _

pub const RIGHT_HAND_Y_UP: Self = _

pub const RIGHT_HAND_Y_DOWN: Self = _

pub const RIGHT_HAND_Z_UP: Self = _

pub const RIGHT_HAND_Z_DOWN: Self = _

pub const LEFT_HAND_X_UP: Self = _

pub const LEFT_HAND_X_DOWN: Self = _

pub const LEFT_HAND_Y_UP: Self = _

pub const LEFT_HAND_Y_DOWN: Self = _

pub const LEFT_HAND_Z_UP: Self = _

pub const LEFT_HAND_Z_DOWN: Self = _