Module re_types::datatypes

The low-level datatypes that components are built from.

They all implement the Datatype trait.

Modules

angle ğŸ”’

angle_ext ğŸ”’

annotation_info ğŸ”’

annotation_info_ext ğŸ”’

class_description ğŸ”’

class_description_ext ğŸ”’

class_description_map_elem ğŸ”’

class_description_map_elem_ext ğŸ”’

class_id ğŸ”’

class_id_ext ğŸ”’

float32 ğŸ”’

keypoint_id ğŸ”’

keypoint_id_ext ğŸ”’

keypoint_pair ğŸ”’

keypoint_pair_ext ğŸ”’

mat3x3 ğŸ”’

mat3x3_ext ğŸ”’

mat4x4 ğŸ”’

material ğŸ”’

material_ext ğŸ”’

mesh_properties ğŸ”’

mesh_properties_ext ğŸ”’

quaternion ğŸ”’

quaternion_ext ğŸ”’

rgba32 ğŸ”’

rgba32_ext ğŸ”’

rotation3d ğŸ”’

rotation3d_ext ğŸ”’

rotation_axis_angle ğŸ”’

rotation_axis_angle_ext ğŸ”’

scale3d ğŸ”’

scale3d_ext ğŸ”’

tensor_buffer ğŸ”’

tensor_buffer_ext ğŸ”’

tensor_data ğŸ”’

tensor_data_ext ğŸ”’

tensor_dimension ğŸ”’

tensor_dimension_ext ğŸ”’

transform3d ğŸ”’

transform3d_ext ğŸ”’

translation_and_mat3x3 ğŸ”’

translation_and_mat3x3_ext ğŸ”’

translation_rotation_scale3d ğŸ”’

translation_rotation_scale3d_ext ğŸ”’

utf8 ğŸ”’

utf8_ext ğŸ”’

uvec2d ğŸ”’

uvec2d_ext ğŸ”’

uvec3d ğŸ”’

uvec3d_ext ğŸ”’

uvec4d ğŸ”’

uvec4d_ext ğŸ”’

vec2d ğŸ”’

vec2d_ext ğŸ”’

vec3d ğŸ”’

vec3d_ext ğŸ”’

vec4d ğŸ”’

vec4d_ext ğŸ”’

Structs

AnnotationInfo

Datatype: Annotation info annotating a class id or key-point id.

ClassDescription

Datatype: The description of a semantic Class.

ClassDescriptionMapElem

Datatype: A helper type for mapping class IDs to class descriptions.

ClassId

Datatype: A 16-bit ID representing a type of semantic class.

Float32

Datatype: A single-precision 32-bit IEEE 754 floating point number.

KeypointId

Datatype: A 16-bit ID representing a type of semantic keypoint within a class.

KeypointPair

Datatype: A connection between two Keypoints.

Mat3x3

Datatype: A 3x3 Matrix.

Mat4x4

Datatype: A 4x4 Matrix.

Material

Datatype: Material properties of a mesh.

MeshProperties

Datatype: Optional triangle indices for a mesh.

Quaternion

Datatype: A Quaternion represented by 4 real numbers.

Rgba32

Datatype: An RGBA color with unmultiplied/separate alpha, in sRGB gamma space with linear alpha.

RotationAxisAngle

Datatype: 3D rotation represented by a rotation around a given axis.

TensorData

Datatype: A multi-dimensional Tensor of data.

TensorDimension

Datatype: A single dimension within a multi-dimensional tensor.

TranslationAndMat3x3

Datatype: Representation of an affine transform via a 3x3 affine matrix paired with a translation.

TranslationRotationScale3D

Datatype: Representation of an affine transform via separate translation, rotation & scale.

UVec2D

Datatype: A uint32 vector in 2D space.

UVec3D

Datatype: A uint32 vector in 3D space.

UVec4D

Datatype: A uint vector in 4D space.

Utf8

Datatype: A string of text, encoded as UTF-8.

Vec2D

Datatype: A vector in 2D space.

Vec3D

Datatype: A vector in 3D space.

Vec4D

Datatype: A vector in 4D space.

Enums

Angle

Datatype: Angle in either radians or degrees.

Rotation3D

Datatype: A 3D rotation.

Scale3D

Datatype: 3D scaling factor, part of a transform representation.

TensorBuffer

Datatype: The underlying storage for a Tensor.

Transform3D

Datatype: Representation of a 3D affine transform.