Struct re_types::datatypes::TensorData

pub struct TensorData {
    pub shape: Vec<TensorDimension>,
    pub buffer: TensorBuffer,
}

Datatype: A multi-dimensional Tensor of data.

The number of dimensions and their respective lengths is specified by the shape field. The dimensions are ordered from outermost to innermost. For example, in the common case of a 2D RGB Image, the shape would be [height, width, channel].

These dimensions are combined with an index to look up values from the buffer field, which stores a contiguous array of typed values.

Fields
shape: Vec<TensorDimension>buffer: TensorBuffer
Implementations

impl TensorData

pub fn new(shape: Vec<TensorDimension>, buffer: TensorBuffer) -> Self

pub fn shape(&self) -> &[TensorDimension]

pub fn shape_short(&self) -> &[TensorDimension]

Returns the shape of the tensor with all trailing dimensions of size 1 ignored.

If all dimension sizes are one, this returns only the first dimension.

pub fn num_dim(&self) -> usize

pub fn image_height_width_channels(&self) -> Option<[u64; 3]>

If the tensor can be interpreted as an image, return the height, width, and channels/depth of it.

pub fn is_shaped_like_an_image(&self) -> bool

Returns true if the tensor can be interpreted as an image.

pub fn is_vector(&self) -> bool

Returns true if either all dimensions have size 1 or only a single dimension has a size larger than 1.

Empty tensors return false.

pub fn get_with_image_coords( &self, x: u64, y: u64, channel: u64 ) -> Option<TensorElement>

Query with x, y, channel indices.

Allows to query values for any image like tensor even if it has more or less dimensions than 3. (useful for sampling e.g. N x M x C x 1 tensor which is a valid image)

pub fn get(&self, index: &[u64]) -> Option<TensorElement>

pub fn dtype(&self) -> TensorDataType

pub fn size_in_bytes(&self) -> usize

impl TensorData

pub fn from_image_file(path: &Path) -> Result<Self, TensorImageLoadError>

Construct a tensor from the contents of an image file on disk.

JPEGs will be kept encoded, left to the viewer to decode on-the-fly. Other images types will be decoded directly.

Requires the image feature.

pub fn from_jpeg_file(path: &Path) -> Result<Self, TensorImageLoadError>

Construct a tensor from the contents of a JPEG file on disk.

Requires the image feature.

pub fn tensor_from_jpeg_file( image_path: impl AsRef<Path> ) -> Result<Self, TensorImageLoadError>

👎Deprecated: Renamed ‘from_jpeg_file’

pub fn from_image_bytes( bytes: Vec<u8>, format: ImageFormat ) -> Result<Self, TensorImageLoadError>

Construct a tensor from the contents of an image file.

JPEGs will be kept encoded, left to the viewer to decode on-the-fly. Other images types will be decoded directly.

Requires the image feature.

pub fn from_jpeg_bytes( jpeg_bytes: Vec<u8> ) -> Result<Self, TensorImageLoadError>

Construct a tensor from the contents of a JPEG file, without decoding it now.

Requires the image feature.

pub fn tensor_from_jpeg_bytes( jpeg_bytes: Vec<u8> ) -> Result<Self, TensorImageLoadError>

👎Deprecated: Renamed ‘from_jpeg_bytes’

pub fn from_image( image: impl Into<DynamicImage> ) -> Result<TensorData, TensorImageLoadError>

Construct a tensor from something that can be turned into a [image::DynamicImage].

Requires the image feature.

This is a convenience function that calls DecodedTensor::from_image.

pub fn from_dynamic_image( image: DynamicImage ) -> Result<TensorData, TensorImageLoadError>

Construct a tensor from [image::DynamicImage].

Requires the image feature.

This is a convenience function that calls DecodedTensor::from_dynamic_image.

pub fn could_be_dynamic_image(&self) -> bool

Predicts if Self::to_dynamic_image is likely to succeed, without doing anything expensive

pub fn to_dynamic_image(&self) -> Result<DynamicImage, TensorImageSaveError>

Try to convert an image-like tensor into an [image::DynamicImage].