Module re_types::image

Image-related utilities.

Enums

ImageConstructionError

Error returned when trying to interpret a tensor as an image.

ImageConversionError

Errors when converting images from the [image] crate to an crate::archetypes::Image.

ImageKind

The kind of image data, either color, segmentation, or depth image.

ImageLoadError

Errors when loading image files.

YuvMatrixCoefficients

Yuv matrix coefficients that determine how a YUV image is meant to be converted to RGB.

Traits

ImageChannelType

Types that implement this can be used as image channel types.

Functions

blob_and_datatype_from_tensor

Converts it to what is useful for the image API.

find_non_empty_dim_indices

Returns the indices of an appropriate set of dimensions.

rgb_from_yuv

Returns sRGB from YUV color.