re_types::image
Function rgb_from_yuv

pub fn rgb_from_yuv(
    y: u8,
    u: u8,
    v: u8,
    limited_range: bool,
    coefficients: YuvMatrixCoefficients,
) -> [u8; 3]

Returns sRGB from YUV color.

This conversion mirrors the function of the same name in yuv_converter.wgsl

Specifying the color standard should be exposed in the future #3541