Struct re_types::tensor_data::DecodedTensor

pub struct DecodedTensor(TensorData);

A thin wrapper around a TensorData that is guaranteed to not be compressed (never a jpeg).

All clones are shallow, like for TensorData.

Tuple Fields
0: TensorData
Implementations

impl DecodedTensor

pub fn inner(&self) -> &TensorData

pub fn into_inner(self) -> TensorData

impl DecodedTensor

pub fn from_image( image: impl Into<DynamicImage> ) -> Result<DecodedTensor, TensorImageLoadError>

Construct a tensor from something that can be turned into a [image::DynamicImage].

Requires the image feature.

pub fn from_dynamic_image( image: DynamicImage ) -> Result<DecodedTensor, TensorImageLoadError>

Construct a tensor from [image::DynamicImage].

Requires the image feature.

pub fn try_decode( maybe_encoded_tensor: TensorData ) -> Result<Self, TensorImageLoadError>

pub fn decode_jpeg_bytes( jpeg_bytes: &ArrowBuffer<u8>, [expected_height, expected_width, expected_channels]: [u64; 3] ) -> Result<DecodedTensor, TensorImageLoadError>