Trait re_types::Archetype

pub trait Archetype {
    // Required methods
    fn name() -> ArchetypeName;
    fn required_components() -> &'static [ComponentName];
    fn recommended_components() -> &'static [ComponentName];
    fn optional_components() -> &'static [ComponentName];
    fn all_components() -> &'static [ComponentName];
    fn try_to_arrow(&self) -> SerializationResult<Vec<(Field, Box<dyn Array>)>>;
    fn try_from_arrow(
        data: impl IntoIterator<Item = (Field, Box<dyn Array>)>
    ) -> DeserializationResult<Self>
       where Self: Sized;

    // Provided methods
    fn to_arrow(&self) -> Vec<(Field, Box<dyn Array>)> { ... }
    fn from_arrow(
        data: impl IntoIterator<Item = (Field, Box<dyn Array>)>
    ) -> Self
       where Self: Sized { ... }
}

Required Methods

fn name() -> ArchetypeName

The fully-qualified name of this archetype, e.g. rerun.archetypes.Points2D.

fn required_components() -> &'static [ComponentName]

The fully-qualified component names of every component that must be provided by the user when constructing this archetype.

fn recommended_components() -> &'static [ComponentName]

The fully-qualified component names of every component that should be provided by the user when constructing this archetype.

fn optional_components() -> &'static [ComponentName]

The fully-qualified component names of every component that could be provided by the user when constructing this archetype.

fn all_components() -> &'static [ComponentName]

All components including required, recommended, and optional.

fn try_to_arrow(&self) -> SerializationResult<Vec<(Field, Box<dyn Array>)>>

Serializes all non-null Components of this Archetype into Arrow arrays.

This can never fail for Rerun’s built-in archetypes. For the non-fallible version, see Archetype::to_arrow.

fn try_from_arrow( data: impl IntoIterator<Item = (Field, Box<dyn Array>)> ) -> DeserializationResult<Self>where Self: Sized,

Given an iterator of Arrow arrays and their respective field metadata, deserializes them into this archetype.

Arrow arrays that are unknown to this Archetype will simply be ignored and a warning logged to stderr.

For the non-fallible version, see Archetype::from_arrow.

Provided Methods

fn to_arrow(&self) -> Vec<(Field, Box<dyn Array>)>

Serializes all non-null Components of this Archetype into Arrow arrays.

Panics on failure. This can never fail for Rerun’s built-in archetypes.

For the fallible version, see Archetype::try_to_arrow.

fn from_arrow(data: impl IntoIterator<Item = (Field, Box<dyn Array>)>) -> Selfwhere Self: Sized,

Given an iterator of Arrow arrays and their respective field metadata, deserializes them into this archetype.

Panics on failure. For the fallible version, see Archetype::try_from_arrow.

Arrow arrays that are unknown to this Archetype will simply be ignored and a warning logged to stderr.