Trait re_types::Loggable

pub trait Loggable: Sized {
    type Name;
    type Item<'a>;
    type Iter<'a>: Iterator<Item = Self::Item<'a>>;

    // Required methods
    fn name() -> Self::Name;
    fn arrow_datatype() -> DataType;
    fn try_to_arrow_opt<'a>(
        data: impl IntoIterator<Item = Option<impl Into<Cow<'a, Self>>>>,
        extension_wrapper: Option<&str>
    ) -> SerializationResult<Box<dyn Array>>
       where Self: Clone + 'a;
    fn try_iter_from_arrow(
        data: &dyn Array
    ) -> DeserializationResult<Self::Iter<'_>>;
    fn convert_item_to_opt_self(item: Self::Item<'_>) -> Option<Self>;

    // Provided methods
    fn to_arrow<'a>(
        data: impl IntoIterator<Item = impl Into<Cow<'a, Self>>>,
        extension_wrapper: Option<&str>
    ) -> Box<dyn Array>
       where Self: Clone + 'a { ... }
    fn try_to_arrow<'a>(
        data: impl IntoIterator<Item = impl Into<Cow<'a, Self>>>,
        extension_wrapper: Option<&str>
    ) -> SerializationResult<Box<dyn Array>>
       where Self: Clone + 'a { ... }
    fn to_arrow_opt<'a>(
        data: impl IntoIterator<Item = Option<impl Into<Cow<'a, Self>>>>,
        extension_wrapper: Option<&str>
    ) -> Box<dyn Array>
       where Self: Clone + 'a { ... }
    fn from_arrow(data: &dyn Array) -> Vec<Self> { ... }
    fn try_from_arrow(data: &dyn Array) -> DeserializationResult<Vec<Self>> { ... }
    fn from_arrow_opt(data: &dyn Array) -> Vec<Option<Self>> { ... }
    fn try_from_arrow_opt(
        data: &dyn Array
    ) -> DeserializationResult<Vec<Option<Self>>> { ... }
    fn convert_item_to_self(item: Self::Item<'_>) -> Self { ... }

}

Anything that can be serialized to and deserialized from Arrow data.

Required Associated Types

type Name

type Item<'a>

type Iter<'a>: Iterator<Item = Self::Item<'a>>

Required Methods

fn name() -> Self::Name

The fully-qualified name of this loggable, e.g. rerun.datatypes.Vec2D.

fn arrow_datatype() -> DataType

The underlying [arrow2::datatypes::DataType].

fn try_to_arrow_opt<'a>( data: impl IntoIterator<Item = Option<impl Into<Cow<'a, Self>>>>, extension_wrapper: Option<&str> ) -> SerializationResult<Box<dyn Array>>where Self: Clone + 'a,

Given an iterator of options of owned or reference values to the current Loggable, serializes them into an Arrow array. The Arrow array’s datatype will match Loggable::arrow_datatype.

This will never fail for Rerun’s built-in Loggable. For the non-fallible version, see Loggable::to_arrow_opt.

fn try_iter_from_arrow( data: &dyn Array ) -> DeserializationResult<Self::Iter<'_>>

Given an Arrow array, deserializes it into a iterator of Loggable::Items.

Note: mostly for reasons related to typing of trait implementations, the implementor of Loggable may choose an arbitrary iterable Loggable::Item that differs from the Loggable itself.

These items can be be converted to an optional Loggable using Loggable::convert_item_to_self.

This is the base deserialization mechanism that all Loggable implementors must provide. All other conversions above can be generated from this primitive.

This will never fail for if the Arrow array’s datatype matches the one returned by Loggable::arrow_datatype.

fn convert_item_to_opt_self(item: Self::Item<'_>) -> Option<Self>

Convert a Loggable::Item into an optional Loggable

This is intended to be used with Loggable::try_iter_from_arrow

Provided Methods

fn to_arrow<'a>( data: impl IntoIterator<Item = impl Into<Cow<'a, Self>>>, extension_wrapper: Option<&str> ) -> Box<dyn Array>where Self: Clone + 'a,

Given an iterator of owned or reference values to the current Loggable, serializes them into an Arrow array. The Arrow array’s datatype will match Loggable::arrow_datatype.

Panics on failure. This will never fail for Rerun’s built-in Loggables.

For the fallible version, see Loggable::try_to_arrow.

fn try_to_arrow<'a>( data: impl IntoIterator<Item = impl Into<Cow<'a, Self>>>, extension_wrapper: Option<&str> ) -> SerializationResult<Box<dyn Array>>where Self: Clone + 'a,

Given an iterator of owned or reference values to the current Loggable, serializes them into an Arrow array. The Arrow array’s datatype will match Loggable::arrow_datatype.

This will never fail for Rerun’s built-in Loggable. For the non-fallible version, see Loggable::to_arrow.

fn to_arrow_opt<'a>( data: impl IntoIterator<Item = Option<impl Into<Cow<'a, Self>>>>, extension_wrapper: Option<&str> ) -> Box<dyn Array>where Self: Clone + 'a,

Given an iterator of options of owned or reference values to the current Loggable, serializes them into an Arrow array. The Arrow array’s datatype will match Loggable::arrow_datatype.

Panics on failure. This will never fail for Rerun’s built-in Loggable.

For the fallible version, see Loggable::try_to_arrow_opt.

fn from_arrow(data: &dyn Array) -> Vec<Self>

Given an Arrow array, deserializes it into a collection of Loggables.

Panics if the data schema doesn’t match, or if optional entries were missing at runtime. For the non-fallible version, see Loggable::try_from_arrow.

fn try_from_arrow(data: &dyn Array) -> DeserializationResult<Vec<Self>>

Given an Arrow array, deserializes it into a collection of Loggables.

This will never fail if the Arrow array’s datatype matches the one returned by Loggable::arrow_datatype. For the non-fallible version, see Loggable::from_arrow_opt.

fn from_arrow_opt(data: &dyn Array) -> Vec<Option<Self>>

Given an Arrow array, deserializes it into a collection of optional Loggables.

This will never fail if the Arrow array’s datatype matches the one returned by Loggable::arrow_datatype. For the fallible version, see Loggable::try_from_arrow_opt.

fn try_from_arrow_opt( data: &dyn Array ) -> DeserializationResult<Vec<Option<Self>>>

Given an Arrow array, deserializes it into a collection of optional Loggables.

This will never fail if the Arrow array’s datatype matches the one returned by Loggable::arrow_datatype. For the non-fallible version, see Loggable::from_arrow_opt.

fn convert_item_to_self(item: Self::Item<'_>) -> Self

Convert a Loggable::Item into a Loggable

This is intended to be used with Loggable::try_iter_from_arrow when the type is known to be non-nullible.