re_types::blueprint
Module archetypes

The blueprint-specific archetypes.

Modules

background 🔒

container_blueprint 🔒

dataframe_query 🔒

entity_behavior 🔒

force_center 🔒

force_collision_radius 🔒

force_link 🔒

force_many_body 🔒

force_position 🔒

line_grid3d 🔒

map_background 🔒

map_zoom 🔒

near_clip_plane 🔒

panel_blueprint 🔒

plot_legend 🔒

scalar_axis 🔒

tensor_scalar_mapping 🔒

tensor_slice_selection 🔒

tensor_view_fit 🔒

view_blueprint 🔒

view_contents 🔒

viewport_blueprint 🔒

visible_time_ranges 🔒

visual_bounds2d 🔒

visualizer_overrides 🔒

Structs

Background

Archetype: Configuration for the background of a view.

ContainerBlueprint

Archetype: The description of a container.

DataframeQuery

Archetype: The query for the dataframe view.

EntityBehavior

Archetype: General visualization behavior of an entity.

ForceCenter

Archetype: Tries to move the center of mass of the graph to the origin.

ForceCollisionRadius

Archetype: Resolves collisions between the bounding circles, according to the radius of the nodes.

ForceLink

Archetype: Aims to achieve a target distance between two nodes that are connected by an edge.

ForceManyBody

Archetype: A force between each pair of nodes that ressembles an electrical charge.

ForcePosition

Archetype: Similar to gravity, this force pulls nodes towards a specific position.

LineGrid3D

Archetype: Configuration for the 3D line grid.

MapBackground

Archetype: Configuration for the background map of the map view.

MapZoom

Archetype: Configuration of the map view zoom level.

NearClipPlane

Archetype: Controls the distance to the near clip plane in 3D scene units.

PanelBlueprint

Archetype: Shared state for the 3 collapsible panels.

PlotLegend

Archetype: Configuration for the legend of a plot.

ScalarAxis

Archetype: Configuration for the scalar axis of a plot.

TensorScalarMapping

Archetype: Configures how tensor scalars are mapped to color.

TensorSliceSelection

Archetype: Specifies a 2D slice of a tensor.

TensorViewFit

Archetype: Configures how a selected tensor slice is shown on screen.

ViewBlueprint

Archetype: The description of a single view.

ViewContents

Archetype: The contents of a View.

ViewportBlueprint

Archetype: The top-level description of the viewport.

VisibleTimeRanges

Archetype: Configures what range of each timeline is shown on a view.

VisualBounds2D

Archetype: Controls the visual bounds of a 2D view.

VisualizerOverrides

Archetype: Override the visualizers for an entity.