re_types::blueprint
Module components

The blueprint-specific components.

Modules

active_tab 🔒

apply_latest_at 🔒

auto_layout 🔒

auto_layout_ext 🔒

auto_views 🔒

background_kind 🔒

column_share 🔒

component_column_selector 🔒

component_column_selector_ext 🔒

container_kind 🔒

corner2d 🔒

corner2d_ext 🔒

enabled 🔒

enabled_ext 🔒

filter_by_range 🔒

filter_by_range_ext 🔒

filter_is_not_null 🔒

filter_is_not_null_ext 🔒

force_distance 🔒

force_iterations 🔒

force_strength 🔒

grid_columns 🔒

grid_spacing 🔒

grid_spacing_ext 🔒

included_content 🔒

lock_range_during_zoom 🔒

map_provider 🔒

near_clip_plane 🔒

near_clip_plane_ext 🔒

panel_state 🔒

panel_state_ext 🔒

query_expression 🔒

root_container 🔒

row_share 🔒

selected_columns 🔒

tensor_dimension_index_slider 🔒

tensor_dimension_index_slider_ext 🔒

timeline_name 🔒

timeline_name_ext 🔒

view_class 🔒

view_class_ext 🔒

view_fit 🔒

view_maximized 🔒

view_origin 🔒

view_origin_ext 🔒

viewer_recommendation_hash 🔒

viewer_recommendation_hash_ext 🔒

visible_time_range 🔒

visual_bounds2d 🔒

visual_bounds2d_ext 🔒

visualizer_override 🔒

zoom_level 🔒

Structs

ActiveTab

Component: The active tab in a tabbed container.

ApplyLatestAt

Component: Whether empty cells in a dataframe should be filled with a latest-at query.

AutoLayout

Component: Whether the viewport layout is determined automatically.

AutoViews

Component: Whether or not views should be created automatically.

ColumnShare

Component: The layout share of a column in the container.

ComponentColumnSelector

Component: Describe a component column to be selected in the dataframe view.

Enabled

Component: Whether a procedure is enabled.

FilterByRange

Component: Configuration for a filter-by-range feature of the dataframe view.

FilterIsNotNull

Component: Configuration for the filter is not null feature of the dataframe view.

ForceDistance

Component: The target distance between two nodes.

ForceIterations

Component: Specifies how often this force should be applied per iteration.

ForceStrength

Component: The strength of a given force.

GridColumns

Component: How many columns a grid container should have.

GridSpacing

Component: Space between grid lines of one line to the next in scene units.

IncludedContent

Component: All the contents in the container.

LockRangeDuringZoom

Component: Indicate whether the range should be locked when zooming in on the data.

NearClipPlane

Component: Distance to the near clip plane used for Spatial2DView.

QueryExpression

Component: An individual query expression used to filter a set of datatypes::EntityPaths.

RootContainer

Component: The container that sits at the root of a viewport.

RowShare

Component: The layout share of a row in the container.

SelectedColumns

Component: Describe a component column to be selected in the dataframe view.

TensorDimensionIndexSlider

Component: Show a slider for the index of some dimension of a slider.

TimelineName

Component: A timeline identified by its name.

ViewClass

Component: The class identifier of view, e.g. "2D", "TextLog", ….

ViewMaximized

Component: Whether a view is maximized.

ViewOrigin

Component: The origin of a view.

ViewerRecommendationHash

Component: Hash of a viewer recommendation.

VisibleTimeRange

Component: The range of values on a given timeline that will be included in a view’s query.

VisualBounds2D

Component: Visual bounds in 2D space used for Spatial2DView.

VisualizerOverride

Component: Single visualizer override the visualizers for an entity.

ZoomLevel

Component: A zoom level determines how much of the world is visible on a map.

Enums

BackgroundKind

Component: The type of the background in a view.

ContainerKind

Component: The kind of a blueprint container (tabs, grid, …).

Corner2D

Component: One of four 2D corners, typically used to align objects.

MapProvider

Component: Name of the map provider to be used in Map views.

PanelState

Component: Tri-state for panel controls.

ViewFit

Component: Determines whether an image or texture should be scaled to fit the viewport.