re_types
Module components

Components are the basic building blocks of archetypes.

They all implement the Component trait.

Each component is a wrapper around a datatype.

Modules

_components 🔒

Structs

AlbedoFactor

Component: A color multiplier, usually applied to a whole entity, e.g. a mesh.

AnnotationContext

Component: The annotation context provides additional information on how to display entities.

AxisLength

Component: The length of an axis in local units of the space.

Blob

Component: A binary blob of data.

ClassId

Component: A 16-bit ID representing a type of semantic class.

ClearIsRecursive

Component: Configures how a clear operation should behave - recursive or not.

Color

Component: An RGBA color with unmultiplied/separate alpha, in sRGB gamma space with linear alpha.

DepthMeter

Component: The world->depth map scaling factor.

DrawOrder

Component: Draw order of 2D elements. Higher values are drawn on top of lower values.

EntityPath

Component: A path to an entity, usually to reference some data that is part of the target entity.

FillRatio

Component: How much a primitive fills out the available space.

GammaCorrection

Component: A gamma correction value to be used with a scalar value or color.

GeoLineString

Component: A geospatial line string expressed in EPSG:4326 latitude and longitude (North/East-positive degrees).

GraphEdge

Component: An edge in a graph connecting two nodes.

GraphNode

Component: A string-based ID representing a node in a graph.

HalfSize2D

Component: Half-size (radius) of a 2D box.

HalfSize3D

Component: Half-size (radius) of a 3D box.

ImageBuffer

Component: A buffer that is known to store image data.

ImageFormat

Component: The metadata describing the contents of a components::ImageBuffer.

ImagePlaneDistance

Component: The distance from the camera origin to the image plane when the projection is shown in a 3D viewer.

Interactive

Component: Whether the entity can be interacted with.

KeypointId

Component: A 16-bit ID representing a type of semantic keypoint within a class.

LatLon

Component: A geospatial position expressed in EPSG:4326 latitude and longitude (North/East-positive degrees).

Length

Component: Length, or one-dimensional size.

LineStrip2D

Component: A line strip in 2D space.

LineStrip3D

Component: A line strip in 3D space.

MarkerSize

Component: Radius of a marker of a point in e.g. a 2D plot, measured in UI points.

MediaType

Component: A standardized media type (RFC2046, formerly known as MIME types), encoded as a string.

Name

Component: A display name, typically for an entity or a item like a plot series.

Opacity

Component: Degree of transparency ranging from 0.0 (fully transparent) to 1.0 (fully opaque).

PinholeProjection

Component: Camera projection, from image coordinates to view coordinates.

Plane3D

Component: An infinite 3D plane represented by a unit normal vector and a distance.

PoseRotationAxisAngle

Component: 3D rotation represented by a rotation around a given axis that doesn’t propagate in the transform hierarchy.

PoseRotationQuat

Component: A 3D rotation expressed as a quaternion that doesn’t propagate in the transform hierarchy.

PoseScale3D

Component: A 3D scale factor that doesn’t propagate in the transform hierarchy.

PoseTransformMat3x3

Component: A 3x3 transformation matrix Matrix that doesn’t propagate in the transform hierarchy.

PoseTranslation3D

Component: A translation vector in 3D space that doesn’t propagate in the transform hierarchy.

Position2D

Component: A position in 2D space.

Position3D

Component: A position in 3D space.

Radius

Component: The radius of something, e.g. a point.

Range1D

Component: A 1D range, specifying a lower and upper bound.

Resolution

Component: Pixel resolution width & height, e.g. of a camera sensor.

RotationAxisAngle

Component: 3D rotation represented by a rotation around a given axis.

RotationQuat

Component: A 3D rotation expressed as a quaternion.

Scalar

Component: A scalar value, encoded as a 64-bit floating point.

Scale3D

Component: A 3D scale factor.

SeriesVisible

Component: Like components::Visible, but for time series.

ShowLabels

Component: Whether the entity’s components::Text label is shown.

StrokeWidth

Component: The width of a stroke specified in UI points.

TensorData

Component: An N-dimensional array of numbers.

TensorDimensionIndexSelection

Component: Specifies a concrete index on a tensor dimension.

TensorHeightDimension

Component: Specifies which dimension to use for height.

TensorWidthDimension

Component: Specifies which dimension to use for width.

Texcoord2D

Component: A 2D texture UV coordinate.

Text

Component: A string of text, e.g. for labels and text documents.

TextLogLevel

Component: The severity level of a text log message.

Timestamp

Component: When the recording started.

TransformMat3x3

Component: A 3x3 transformation matrix Matrix.

Translation3D

Component: A translation vector in 3D space.

TriangleIndices

Component: The three indices of a triangle in a triangle mesh.

ValueRange

Component: Range of expected or valid values, specifying a lower and upper bound.

Vector2D

Component: A vector in 2D space.

Vector3D

Component: A vector in 3D space.

VideoTimestamp

Component: Timestamp inside a archetypes::AssetVideo.

ViewCoordinates

Component: How we interpret the coordinate system of an entity/space.

Visible

Component: Whether the container, view, entity or instance is currently visible.

Enums

AggregationPolicy

Component: Policy for aggregation of multiple scalar plot values.

Colormap

Component: Colormap for mapping scalar values within a given range to a color.

FillMode

Component: How a geometric shape is drawn and colored.

GraphType

Component: Specifies if a graph has directed or undirected edges.

MagnificationFilter

Component: Filter used when magnifying an image/texture such that a single pixel/texel is displayed as multiple pixels on screen.

MarkerShape

Component: The visual appearance of a point in e.g. a 2D plot.

TransformRelation

Component: Specifies relation a spatial transform describes.