Module re_types::components

Components are the basic building blocks of archetypes.

They all implement the Component trait.

Each component is a wrapper around a datatype.

Modules

annotation_context 🔒

blob 🔒

blob_ext 🔒

class_id 🔒

class_id_ext 🔒

clear_is_recursive 🔒

color 🔒

color_ext 🔒

depth_meter 🔒

depth_meter_ext 🔒

disconnected_space 🔒

disconnected_space_ext 🔒

draw_order 🔒

draw_order_ext 🔒

half_sizes2d 🔒

half_sizes2d_ext 🔒

half_sizes3d 🔒

half_sizes3d_ext 🔒

instance_key 🔒

instance_key_ext 🔒

keypoint_id 🔒

keypoint_id_ext 🔒

line_strip2d 🔒

line_strip2d_ext 🔒

line_strip3d 🔒

line_strip3d_ext 🔒

material 🔒

material_ext 🔒

media_type 🔒

media_type_ext 🔒

mesh_properties 🔒

mesh_properties_ext 🔒

out_of_tree_transform3d 🔒

pinhole_projection 🔒

pinhole_projection_ext 🔒

position2d 🔒

position2d_ext 🔒

position3d 🔒

position3d_ext 🔒

radius 🔒

radius_ext 🔒

resolution 🔒

rotation3d 🔒

rotation3d_ext 🔒

scalar 🔒

scalar_ext 🔒

scalar_scattering 🔒

tensor_data 🔒

text 🔒

text_ext 🔒

text_log_level 🔒

text_log_level_ext 🔒

transform3d 🔒

transform3d_ext 🔒

vector3d 🔒

vector3d_ext 🔒

view_coordinates 🔒

view_coordinates_ext 🔒

Structs

AnnotationContext

Component: The AnnotationContext provides additional information on how to display entities.

Blob

Component: A binary blob of data.

ClassId

Component: A 16-bit ID representing a type of semantic class.

ClearIsRecursive

Component: Configures how a clear operation should behave - recursive or not?

Color

Component: An RGBA color with unmultiplied/separate alpha, in sRGB gamma space with linear alpha.

DepthMeter

Component: A component indicating how long a meter is, expressed in native units.

DisconnectedSpace

Component: Specifies that the entity path at which this is logged is disconnected from its parent.

DrawOrder

Component: Draw order used for the display order of 2D elements.

HalfSizes2D

Component: Half-sizes (extents) of a 2D box along its local axis, starting at its local origin/center.

HalfSizes3D

Component: Half-sizes (extents) of a 3D box along its local axis, starting at its local origin/center.

InstanceKey

Component: A unique numeric identifier for each individual instance within a batch.

KeypointId

Component: A 16-bit ID representing a type of semantic keypoint within a class.

LineStrip2D

Component: A line strip in 2D space.

LineStrip3D

Component: A line strip in 3D space.

Material

Component: Material properties of a mesh.

MediaType

Component: A standardized media type (RFC2046, formerly known as MIME types), encoded as a utf8 string.

MeshProperties

Component: Optional triangle indices for a mesh.

OutOfTreeTransform3D

Component: An out-of-tree affine transform between two 3D spaces, represented in a given direction.

PinholeProjection

Component: Camera projection, from image coordinates to view coordinates.

Position2D

Component: A position in 2D space.

Position3D

Component: A position in 3D space.

Radius

Component: A Radius component.

Resolution

Component: Pixel resolution width & height, e.g. of a camera sensor.

Rotation3D

Component: A 3D rotation, represented either by a quaternion or a rotation around axis.

Scalar

Component: A double-precision scalar.

ScalarScattering

Component: If true, a scalar will be shown as individual point in a scatter plot.

TensorData

Component: A multi-dimensional Tensor with optionally named arguments.

Text

Component: A string of text, e.g. for labels and text documents

TextLogLevel

Component: The severity level of a text log message.

Transform3D

Component: An affine transform between two 3D spaces, represented in a given direction.

Vector3D

Component: A vector in 3D space.

ViewCoordinates

Component: How we interpret the coordinate system of an entity/space.