Module re_types::datatypes

The low-level datatypes that components are built from.

They all implement the Datatype trait.

Modules

_datatypes 🔒

Structs

AnnotationInfo

Datatype: Annotation info annotating a class id or key-point id.

ClassDescription

Datatype: The description of a semantic Class.

ClassDescriptionMapElem

Datatype: A helper type for mapping class IDs to class descriptions.

ClassId

Datatype: A 16-bit ID representing a type of semantic class.

EntityPath

Datatype: A path to an entity in the DataStore.

Float32

Datatype: A single-precision 32-bit IEEE 754 floating point number.

KeypointId

Datatype: A 16-bit ID representing a type of semantic keypoint within a class.

KeypointPair

Datatype: A connection between two Keypoints.

Mat3x3

Datatype: A 3x3 Matrix.

Mat4x4

Datatype: A 4x4 Matrix.

Material

Datatype: Material properties of a mesh.

MeshProperties

Datatype: Optional triangle indices for a mesh.

Quaternion

Datatype: A Quaternion represented by 4 real numbers.

Rgba32

Datatype: An RGBA color with unmultiplied/separate alpha, in sRGB gamma space with linear alpha.

RotationAxisAngle

Datatype: 3D rotation represented by a rotation around a given axis.

TensorData

Datatype: A multi-dimensional Tensor of data.

TensorDimension

Datatype: A single dimension within a multi-dimensional tensor.

TranslationAndMat3x3

Datatype: Representation of an affine transform via a 3x3 affine matrix paired with a translation.

TranslationRotationScale3D

Datatype: Representation of an affine transform via separate translation, rotation & scale.

UInt32

Datatype: A 32bit unsigned integer.

UVec2D

Datatype: A uint32 vector in 2D space.

UVec3D

Datatype: A uint32 vector in 3D space.

UVec4D

Datatype: A uint vector in 4D space.

Utf8

Datatype: A string of text, encoded as UTF-8.

Uuid

Datatype: A 16-byte uuid.

Vec2D

Datatype: A vector in 2D space.

Vec3D

Datatype: A vector in 3D space.

Vec4D

Datatype: A vector in 4D space.

Enums

Angle

Datatype: Angle in either radians or degrees.

Rotation3D

Datatype: A 3D rotation.

Scale3D

Datatype: 3D scaling factor, part of a transform representation.

TensorBuffer

Datatype: The underlying storage for a Tensor.

Transform3D

Datatype: Representation of a 3D affine transform.