Trait re_types::SizeBytes

pub trait SizeBytes {
    // Required method
    fn heap_size_bytes(&self) -> u64;

    // Provided methods
    fn total_size_bytes(&self) -> u64 { ... }
    fn stack_size_bytes(&self) -> u64 { ... }
}

Approximations of stack and heap size for both internal and external types.

Motly used for statistics and triggering events such as garbage collection.

Required Methods

fn heap_size_bytes(&self) -> u64

Returns the total size of self on the heap, in bytes.

Provided Methods

fn total_size_bytes(&self) -> u64

Returns the total size of self in bytes, accounting for both stack and heap space.

fn stack_size_bytes(&self) -> u64

Returns the total size of self on the stack, in bytes.

Defaults to std::mem::size_of_val(self).